refactor(wizard): #1544 stage 5 — extract create_course newCourseScaffoldPath + new-path pin
Stage 5 of 8 (Stage 6 folded in per the brief). Lifts the ~520-LOC
new-course scaffold branch from create_course.ts into two new helpers:

  - tools/create_course/_new-path.ts (399 LOC) — full new-path
    orchestration: Subject create, dedup-by-name (signals orchestrator
    to recurse through Stage 4), scaffoldDomain, #207 placeholder
    unlink + #607 single-primary-subject invariant, config persistence,
    course identity spec overlay, pack-subject linking + #385
    COURSE_REFERENCE bridge + #492 Phase 5 direct PlaybookSource,
    #338 projection, onboarding welcome + behavior targets (Stage 6
    fold), and student-visible media attachment to onboarding flow
    phases.
  - tools/create_course/_new-config-merge.ts (152 LOC) — pure function
    mirroring _reuse-config-merge but also folding the assessment-target
    → ACHIEVE-goal projection and the learning-outcome → LEARN-goal
    projection with #447 rubric-prose guard.

AC's new-path dispatcher pin landed in the same commit:
`tests/lib/chat/wizard-tool-executor-dispatcher.test.ts` now carries
"create_course (new path — #607 invariant)" — asserts
`unlinkNonPrimaryPlaybookSubjects` fires on the fresh-scaffold branch
with the freshly-created Subject id AND that `scaffoldDomain` is
called with `forceNewPlaybook: true` so the path is truly fresh-scaffold.
Mock surface extended for `@/lib/domain/scaffold`,
`@/lib/domain/persona-loaders`, `@/lib/content-trust/teaching-profiles`,
plus prisma models (subject.create/findUniqueOrThrow, subjectDomain,
playbookSubject, playbook.findFirst, domain.update).

Orchestrator imports tightened — `updatePlaybookConfig`,
`updateDomainConfig`, `applyStudentExperienceConfig`, `scaffoldDomain`,
the three persona-loader names, `applyBehaviorTargets`,
`behaviorTargetsFromPresets`, `slugify`, `generateLessonPlan` all moved
into the helper or removed as dead.

Behaviour-preserving relative to the pre-#1544 inline block at
create_course.ts L451-871. The recurse path now goes through an
orchestrator-level `return execute(...)` (one stack frame deeper than
pre-extract, identical semantics).

Net file shape:
- create_course.ts: 887 → 376 LOC (-511)
- new: _new-path.ts (399), _new-config-merge.ts (152)

Tests
- 12/12 dispatcher pins green (graph guard + reuse #607 unlink + NEW
  new-path #607 unlink + update_setup + mark_complete + create_institution)
- 161/161 tests/lib/chat suite green (was 160, +1 from new pin)
- 0 new tsc errors introduced
- AC respected: every file under tools/create_course/ ≤ 400 LOC
  (_new-path.ts at 399, _reuse-path.ts at 393, all others ≤ 152)

Refs #1544
