+++ apps/admin/lib/chat/wizard-tool-executor/tools/create_course.ts
@@ -1,7 +1,4 @@
-import { updatePlaybookConfig } from "@/lib/playbook/update-playbook-config";
-import { updateDomainConfig } from "@/lib/domain/update-domain-config";
 import type { WizardToolExec } from "../_shared/types";
-import { applyStudentExperienceConfig } from "../_shared/apply-student-experience";
 
 export async function execute(
   input: Record<string, unknown>,
@@ -15,13 +12,8 @@ if (graphGuard.earlyReturn) return graphGuard.earlyReturn;
 // Server-side: full course creation with scaffolding (identity spec, playbook, system specs, publish, onboarding)
 try {
   const { prisma } = await import("@/lib/prisma");
-  const { scaffoldDomain } = await import("@/lib/domain/scaffold");
-  const { loadPersonaFlowPhases, loadPersonaArchetype, loadPersonaWelcomeTemplate } = await import("@/lib/domain/persona-loaders");
-  const { applyBehaviorTargets, behaviorTargetsFromPresets } = await import("@/lib/domain/agent-tuning");
   const { enrollCaller } = await import("@/lib/enrollment");
   const { randomFakeName } = await import("@/lib/fake-names");
-  const slugify = (await import("slugify")).default;
-  const { generateLessonPlan } = await import("@/lib/content-trust/lesson-planner");
 
   // ── Stage 3 — subject-discipline guard (extracted #1544) ──
   // Runs before domain resolution so a missing subjectDiscipline never
@@ -65,528 +57,25 @@ try {
   });
   if (reuseResult.ok) return reuseResult.earlyReturn;
 
-  // 1. Create or find Subject (course-scoped slug to prevent content bleeding)
-  //    Each course gets its own subject even if the discipline name is the same.
-  //    e.g. "abacus-academy-pw-secret-garden-1005-english-language"
-  const domainRow = await prisma.domain.findUnique({ where: { id: domainId }, select: { slug: true } });
-  const courseSlug = slugify(courseName, { lower: true, strict: true });
-  const disciplineSlug = slugify(subjectDiscipline, { lower: true, strict: true });
-  const subjectSlug = `${domainRow!.slug}-${courseSlug}-${disciplineSlug}`;
-  // Subject.slug is @unique — findUnique reflects the schema invariant.
-  let subject = await prisma.subject.findUnique({ where: { slug: subjectSlug } });
-  if (!subject) {
-    const { suggestTeachingProfile } = await import("@/lib/content-trust/teaching-profiles");
-    subject = await prisma.subject.create({
-      data: {
-        slug: subjectSlug,
-        name: subjectDiscipline,
-        isActive: true,
-        teachingProfile: suggestTeachingProfile(subjectDiscipline),
-      },
-    });
+  // ── Stage 5 — new-course scaffold (extracted #1544; folds in Stage 6) ──
+  // Owns Subject create, dedup-by-name (recurses through reuse path),
+  // scaffold + #607 invariant, config merge, identity-spec overlay,
+  // pack-subject linking, COURSE_REFERENCE bridge + projection, onboarding
+  // welcome + behaviorTargets, and student-visible media attachment.
+  const { newCourseScaffoldPath } = await import("./create_course/_new-path");
+  const newPathCtx = { input, userId, setupData, domainId, subjectDiscipline, courseName, interactionPattern, packSubjectIds, uploadSourceIds };
+  const newPathResult = await newCourseScaffoldPath(newPathCtx);
+  if (!newPathResult.ok) {
+    return execute(input, userId, newPathResult.recurse.setupData);
   }
-
-  // 2. Link Subject → Domain
-  const existingSubjectLink = await prisma.subjectDomain.findFirst({
-    where: { subjectId: subject.id, domainId },
+  const { playbookId, subjectId: subjectIdFromScaffold, subjectIdsToLink, resolvedWelcome, mediaLookup } = newPathResult.state;
+  // Stages 7+8 use `subject` (the Prisma row shape). Re-fetch to preserve
+  // type compatibility — orchestrator-level cost is one extra DB read,
+  // which the next stage extracts away.
+  const subject = await prisma.subject.findUniqueOrThrow({
+    where: { id: subjectIdFromScaffold },
+    select: { id: true, name: true, slug: true },
   });
-  if (!existingSubjectLink) {
-    await prisma.subjectDomain.create({
-      data: { subjectId: subject.id, domainId },
-    });
-  }
-
-  // 3. Dedup guard: if a playbook with the same name already exists in this domain,
-  //    treat it as the existing course (prevents duplicates from AI retries)
-  const existingDupe = await prisma.playbook.findFirst({
-    where: {
-      domainId,
-      name: { equals: courseName, mode: "insensitive" },
-    },
-    select: { id: true },
-  });
-  if (existingDupe) {
-    console.log(`[wizard-tools] create_course: playbook "${courseName}" already exists in domain ${domainId} — reusing ${existingDupe.id}`);
-    // Re-enter the existing-course path by setting existingPlaybookId
-    // and recursing through the same tool (setupData is immutable here,
-    // so we call ourselves with the draftPlaybookId patched in).
-    return execute(
-      input,
-      userId,
-      { ...setupData, draftPlaybookId: existingDupe.id },
-    );
-  }
-
-  // 4. Resolve archetype + flow phases from interaction pattern
-  const archetypeSlug = await loadPersonaArchetype(interactionPattern);
-  const flowPhases = await loadPersonaFlowPhases(interactionPattern);
-
-  // 5. Scaffold domain (identity spec + playbook + system specs + publish + onboarding)
-  const groupId = (input.groupId as string) || (setupData?.groupId as string) || undefined;
-  const scaffoldResult = await scaffoldDomain(domainId, {
-    extendsAgent: archetypeSlug || undefined,
-    flowPhases: flowPhases || undefined,
-    forceNewPlaybook: true,
-    playbookName: courseName,
-    groupId,
-  });
-
-  if (!scaffoldResult.playbook) {
-    throw new Error("Scaffold failed to create playbook");
-  }
-
-  const playbookId = scaffoldResult.playbook.id;
-
-  // 4b. Link primary subject to playbook (step 7 only links pack subjects)
-  await prisma.playbookSubject.upsert({
-    where: { playbookId_subjectId: { playbookId, subjectId: subject.id } },
-    update: {},
-    create: { playbookId, subjectId: subject.id },
-  });
-
-  // Remove any placeholder PlaybookSubjects accumulated on this playbook
-  // from earlier wizard turns (e.g. a Subject named "Course" created when
-  // subjectDiscipline fell back to courseName). See #207.
-  //
-  // Then enforce the single-primary-subject invariant — quick-launch/analyze
-  // (or any other upstream caller) may have linked a domain-level Subject
-  // with a different (non-placeholder) name that we must displace so the
-  // composed prompt doesn't render two CONTENT AUTHORITY sections. See #607.
-  const { removePlaceholderPlaybookSubjects, unlinkNonPrimaryPlaybookSubjects } = await import(
-    "@/lib/knowledge/cleanup-placeholder-subjects"
-  );
-  await removePlaceholderPlaybookSubjects(playbookId, subject.id);
-  const unlink = await unlinkNonPrimaryPlaybookSubjects(playbookId, subject.id);
-  if (unlink.removed > 0) {
-    console.log(
-      `[wizard-tools] create_course: displaced ${unlink.removed} non-primary PlaybookSubject(s) on playbook ${playbookId}: ${unlink.displaced
-        .map((d) => `"${d.subjectName}"`)
-        .join(", ")}`,
-    );
-  }
-
-  // Dual-write: sync PlaybookSource from primary subject
-  // Skip when uploadSourceIds provided — Phase 5 (step 7c) creates PlaybookSource
-  // directly, and syncPlaybookSources would pull in ALL sources for this subject.
-  if (!uploadSourceIds?.length) {
-    const { syncPlaybookSources } = await import("@/lib/knowledge/domain-sources");
-    await syncPlaybookSources(playbookId, subject.id);
-  }
-
-  // 5. Store config in playbook
-  // Fall back to setupData (wizard data bag) for fields the AI may not repeat in create_course
-  const pb = await prisma.playbook.findUnique({
-    where: { id: playbookId },
-    select: { config: true },
-  });
-  const existingConfig = (pb?.config as Record<string, unknown>) || {};
-  const configUpdate: Record<string, unknown> = { ...existingConfig };
-  if (interactionPattern) configUpdate.interactionPattern = interactionPattern;
-  const newTeachingMode = (input.teachingMode as string) || (setupData?.teachingMode as string);
-  if (newTeachingMode) configUpdate.teachingMode = newTeachingMode;
-  if (subjectDiscipline) configUpdate.subjectDiscipline = subjectDiscipline;
-  const newWelcomeMessage = (input.welcomeMessage as string) || (setupData?.welcomeMessage as string);
-  if (newWelcomeMessage) configUpdate.welcomeMessage = newWelcomeMessage;
-  const newSessionCount = input.sessionCount ?? setupData?.sessionCount;
-  if (newSessionCount) configUpdate.sessionCount = Number(newSessionCount);
-  const newDurationMins = input.durationMins ?? setupData?.durationMins;
-  if (newDurationMins) configUpdate.durationMins = Number(newDurationMins);
-  const newPlanEmphasis = (input.planEmphasis as string) || (setupData?.planEmphasis as string);
-  if (newPlanEmphasis) configUpdate.planEmphasis = newPlanEmphasis;
-  const newAudience = (input.audience as string) || (setupData?.audience as string);
-  if (newAudience) configUpdate.audience = newAudience;
-  const newLessonPlanModel = (input.lessonPlanModel as string) || (setupData?.lessonPlanModel as string);
-  if (newLessonPlanModel) configUpdate.lessonPlanModel = newLessonPlanModel;
-
-  // #253 follow-up: progressionMode → modulesAuthored mirror.
-  // The existing-path branch above (line ~796) already does this, but
-  // the new-path branch was missing it — net effect: brand-new courses
-  // landed with modulesAuthored=null, surfacing a "Mode not set" pill on
-  // the course page even when the wizard had collected a clear answer.
-  const newProgressionMode =
-    (input.progressionMode as string) || (setupData?.progressionMode as string);
-  if (newProgressionMode === "learner-picks") {
-    configUpdate.modulesAuthored = true;
-  } else if (newProgressionMode === "ai-led") {
-    configUpdate.modulesAuthored = false;
-  }
-  const newPhysicalMaterials = (input.physicalMaterials as string) || (setupData?.physicalMaterials as string);
-  if (newPhysicalMaterials) configUpdate.physicalMaterials = newPhysicalMaterials;
-  const newCourseContext = (input.courseContext as string) || (setupData?.courseContext as string);
-  if (newCourseContext) configUpdate.courseContext = newCourseContext;
-  const newConstraints = (input.constraints as string[]) || (setupData?.constraints as string[]);
-  if (newConstraints) configUpdate.constraints = newConstraints;
-  // #167 — Carry through pedagogy detected from an uploaded course
-  // reference. Mirrors the existing-playbook path above.
-  const newPedagogy = setupData?.coursePedagogy as {
-    lessonPlanMode?: "structured" | "continuous" | null;
-    cadenceMinutesPerCall?: number | null;
-    suggestedSessionCount?: number | null;
-  } | undefined;
-  if (newPedagogy?.lessonPlanMode) {
-    configUpdate.lessonPlanMode = newPedagogy.lessonPlanMode;
-  }
-  if (newPedagogy?.cadenceMinutesPerCall && !newDurationMins) {
-    configUpdate.durationMins = newPedagogy.cadenceMinutesPerCall;
-  }
-  if (newPedagogy?.suggestedSessionCount && !newSessionCount) {
-    configUpdate.sessionCount = newPedagogy.suggestedSessionCount;
-  }
-
-  // Map assessment targets into goal templates
-  if (input.assessmentTargets) {
-    const existingGoals = (configUpdate.goals as any[]) || [];
-    const newAssessmentGoals = (input.assessmentTargets as string[]).map((t: string) => ({
-      type: "ACHIEVE",
-      name: t,
-      isAssessmentTarget: true,
-      isDefault: true,
-      priority: 8,
-    }));
-    configUpdate.goals = [
-      ...existingGoals.filter((g: any) => !g.isAssessmentTarget),
-      ...newAssessmentGoals,
-    ];
-  }
-  // Map learning outcomes into LEARN goals (from wizard or setupData).
-  // #447 — guarded against AI returning rubric calibration prose (band
-  // descriptors, tier-name lines) as "learning outcomes". The wizard's
-  // bare LEARN templates carry no `ref` and no `sourceContentId`, so
-  // applyProjection() can't diff them away. The guard runs the AI list
-  // through a regex filter + a soft gate that defers entirely to
-  // projection when OUT-NN templates already exist on the playbook.
-  const learningOutcomes = (input.learningOutcomes as string[])
-    || (setupData?.learningOutcomes as string[]);
-  if (learningOutcomes && learningOutcomes.length > 0) {
-    const existingGoals = (configUpdate.goals as any[]) || [];
-    const { guardAILearningOutcomes } = await import("@/lib/chat/wizard-ai-output-guard");
-    const guard = guardAILearningOutcomes(learningOutcomes, existingGoals);
-
-    for (const dropped of guard.filtered) {
-      console.warn(
-        `[wizard:guard] dropped LO "${dropped.value.slice(0, 80)}" — matches rubric pattern ${dropped.pattern}`,
-      );
-    }
-    if (guard.skippedByGate) {
-      console.log(
-        `[wizard:guard] skipped AI learning-outcome extraction — ${guard.gateReason}`,
-      );
-    }
-
-    if (guard.accepted.length > 0) {
-      const existingNames = new Set(existingGoals.map((g: any) => g.name?.toLowerCase().trim()));
-      const newLOGoals = guard.accepted
-        .filter((lo: string) => !existingNames.has(lo.toLowerCase().trim()))
-        .map((lo: string) => ({
-          type: "LEARN",
-          name: lo,
-          isDefault: true,
-          priority: 5,
-        }));
-      if (newLOGoals.length > 0) {
-        configUpdate.goals = [...existingGoals, ...newLOGoals];
-      }
-    }
-  }
-
-  // Student experience config — welcome + sessionFlow.intake mirror + nps
-  applyStudentExperienceConfig(
-    setupData as Record<string, unknown> | undefined,
-    configUpdate,
-    "create_course (new path)",
-    playbookId,
-  );
-
-  // #826 — central helper, skipTimestamp: true because this is
-  // create_course NEW path — the playbook was created in the same
-  // wizard step and has no enrolled callers yet, so no downstream
-  // staleness to mark.
-  await updatePlaybookConfig(
-    playbookId,
-    () => configUpdate,
-    { skipTimestamp: true, reason: "wizard create_course (new path)" },
-  );
-
-  // 6. Link Subject → Playbook
-  await prisma.playbookSubject.upsert({
-    where: { playbookId_subjectId: { playbookId, subjectId: subject.id } },
-    update: {},
-    create: { playbookId, subjectId: subject.id },
-  });
-
-  // Dual-write: sync PlaybookSource from primary subject (idempotent if already done at 4b)
-  // Skip when uploadSourceIds provided — Phase 5 (step 7c) handles it.
-  if (!uploadSourceIds?.length) {
-    const { syncPlaybookSources: syncStep6 } = await import("@/lib/knowledge/domain-sources");
-    await syncStep6(playbookId, subject.id);
-  }
-
-  // 6b. Create per-course identity spec overlay
-  //     Extends the domain identity spec so course-specific teaching rules
-  //     are scoped to this course and don't bleed into other courses.
-  const domainForSpec = await prisma.domain.findUnique({
-    where: { id: domainId },
-    select: { onboardingIdentitySpecId: true },
-  });
-  if (domainForSpec?.onboardingIdentitySpecId) {
-    const domainSpec = await prisma.analysisSpec.findUnique({
-      where: { id: domainForSpec.onboardingIdentitySpecId },
-      select: { slug: true },
-    });
-    if (domainSpec) {
-      const courseIdentitySlug = `${slugify(courseName, { lower: true, strict: true })}-identity`;
-      // #829 — CREATE-only upsert. `update: {}` is a no-op so no
-      // compose-affecting mutation can land here, and the brand-
-      // new spec has no enrolled callers. The linking Playbook
-      // write that happens in the same wizard step is what marks
-      // downstream callers stale once they exist. No helper bump
-      // required at this site.
-      const courseIdentity = await prisma.analysisSpec.upsert({
-        where: { slug: courseIdentitySlug },
-        update: {},
-        create: {
-          slug: courseIdentitySlug,
-          name: `${courseName} Identity`,
-          description: `Course overlay for ${courseName} — extends domain identity with course-specific teaching rules.`,
-          outputType: "COMPOSE",
-          specRole: "IDENTITY",
-          specType: "DOMAIN",
-          scope: "DOMAIN",
-          domain: "identity",
-          isActive: true,
-          isDirty: false,
-          isDeletable: true,
-          extendsAgent: domainSpec.slug,
-          config: { parameters: [] },
-        },
-      });
-      // Link at sortOrder -1 so resolveSpecs picks it before domain overlay (sortOrder 0)
-      const existingLink = await prisma.playbookItem.findFirst({
-        where: { playbookId, specId: courseIdentity.id },
-      });
-      if (!existingLink) {
-        await prisma.playbookItem.create({
-          data: {
-            playbookId,
-            itemType: "SPEC",
-            specId: courseIdentity.id,
-            sortOrder: -1,
-            isEnabled: true,
-          },
-        });
-      }
-    }
-  }
-
-  // 7. Link content-upload subjects from PackUploadStep (if any)
-  //    Only link subjects explicitly passed via packSubjectIds (from the upload step).
-  //    No domain-wide fallback — that caused content from other courses on the
-  //    same domain to bleed into new courses.
-  const subjectIdsToLink = (packSubjectIds ?? [])
-    .filter(id => id !== subject.id); // primary subject already linked at step 4b
-  for (const packSubId of subjectIdsToLink) {
-    await prisma.playbookSubject.upsert({
-      where: { playbookId_subjectId: { playbookId, subjectId: packSubId } },
-      update: {},
-      create: { playbookId, subjectId: packSubId },
-    });
-    // Dual-write: sync PlaybookSource from pack subject
-    // Skip when uploadSourceIds provided — Phase 5 (step 7c) handles it.
-    if (!uploadSourceIds?.length) {
-      const { syncPlaybookSources: syncPackSub } = await import("@/lib/knowledge/domain-sources");
-      await syncPackSub(playbookId, packSubId);
-    }
-
-    const domainLink = await prisma.subjectDomain.findFirst({
-      where: { subjectId: packSubId, domainId },
-    });
-    if (!domainLink) {
-      await prisma.subjectDomain.create({
-        data: { subjectId: packSubId, domainId },
-      });
-    }
-  }
-
-  // 7b. Bridge COURSE_REFERENCE sources to the primary subject.
-  //     LEGACY: Ingest now puts all docs (including pedagogy) on the primary subject.
-  //     Kept for backward compatibility with courses that have fragmented subjects.
-  //     Skip when uploadSourceIds provided — Phase 5 (step 7c) handles all sources directly.
-  if (!uploadSourceIds?.length) {
-    for (const packSubId of subjectIdsToLink) {
-      const packSources = await prisma.subjectSource.findMany({
-        where: { subjectId: packSubId },
-        select: { sourceId: true, source: { select: { documentType: true } } },
-      });
-      for (const ps of packSources) {
-        // #385 Slice 1 Phase 3 — bridge all four COURSE_REFERENCE* values.
-        if (
-          ps.source.documentType === "COURSE_REFERENCE" ||
-          ps.source.documentType === "COURSE_REFERENCE_CANONICAL" ||
-          ps.source.documentType === "COURSE_REFERENCE_TUTOR_BRIEFING" ||
-          ps.source.documentType === "COURSE_REFERENCE_ASSESSOR_RUBRIC" ||
-          ps.source.documentType === "POLICY_DOCUMENT"
-        ) {
-          const existingLink = await prisma.subjectSource.findFirst({
-            where: { subjectId: subject.id, sourceId: ps.sourceId },
-          });
-          if (!existingLink) {
-            await prisma.subjectSource.create({
-              data: { subjectId: subject.id, sourceId: ps.sourceId },
-            });
-          }
-          // Dual-write: PlaybookSource for bridged source
-          const { upsertPlaybookSource: upsertBridgeNew } = await import("@/lib/knowledge/domain-sources");
-          await upsertBridgeNew(playbookId, ps.sourceId);
-        }
-      }
-    }
-  }
-
-  // 7c. Direct PlaybookSource creation from uploadSourceIds (Phase 5)
-  //     When ingest provides sourceIds directly, create PlaybookSource without
-  //     needing the Subject → SubjectSource chain.
-  if (uploadSourceIds?.length) {
-    // Pre-flight FK race guard (incident 2026-05-19, course e5f379ed).
-    const { preflightPlaybookSourceIds, upsertPlaybookSource } =
-      await import("@/lib/knowledge/domain-sources");
-    await preflightPlaybookSourceIds(uploadSourceIds);
-    for (const srcId of uploadSourceIds) {
-      await upsertPlaybookSource(playbookId, srcId);
-    }
-  }
-
-  // 7d. COURSE_REFERENCE projection (#338) — derive CurriculumModule,
-  //     BehaviorTargets, Parameters, and Goal templates from any linked
-  //     COURSE_REFERENCE source. Race-safe (skips sources whose
-  //     extraction isn't ready), best-effort (does not fail
-  //     create_course on projection errors — the projection can be
-  //     re-run later from a re-process button). See
-  //     docs/CONTENT-PIPELINE.md §4 Phase 2.5.
-  try {
-    const { runProjectionForPlaybook } = await import("@/lib/wizard/run-projection-for-playbook");
-    await runProjectionForPlaybook(playbookId);
-  } catch (err) {
-    console.error(
-      `[projection] create_course: projection failed for playbook=${playbookId} — course still created. Error:`,
-      err instanceof Error ? err.message : err,
-    );
-  }
-
-  // 8. Configure onboarding (welcome message + behavior targets)
-  const resolvedWelcome = (input.welcomeMessage as string)
-    || await loadPersonaWelcomeTemplate(interactionPattern)
-    || null;
-
-  const domainUpdate: Record<string, unknown> = {};
-  if (resolvedWelcome) domainUpdate.onboardingWelcome = resolvedWelcome;
-
-  const behaviorTargets = input.behaviorTargets as Record<string, number> | undefined;
-  const personalityPreset = input.personalityPreset as string | undefined;
-  const resolvedTargets =
-    (behaviorTargets && Object.keys(behaviorTargets).length > 0)
-      ? behaviorTargets
-      : personalityPreset
-        ? behaviorTargetsFromPresets(personalityPreset)
-        : undefined;
-  if (resolvedTargets && Object.keys(resolvedTargets).length > 0) {
-    const wrapped: Record<string, { value: number; confidence: number }> = {};
-    for (const [paramId, value] of Object.entries(resolvedTargets)) {
-      wrapped[paramId] = { value, confidence: 0.5 };
-    }
-    domainUpdate.onboardingDefaultTargets = wrapped;
-    await applyBehaviorTargets(playbookId, resolvedTargets);
-  }
-
-  if (Object.keys(domainUpdate).length > 0) {
-    await prisma.domain.update({ where: { id: domainId }, data: domainUpdate });
-  }
-
-  // 8b. Wire student-visible media into onboarding flow phases
-  //     So the AI proactively shares materials during the first call,
-  //     and the educator can see/edit attachments in the First Call Preview.
-  const allSubjectIds = [subject.id, ...subjectIdsToLink];
-  const { isStudentVisibleDefault } = await import("@/lib/doc-type-icons");
-  const studentMedia = await prisma.subjectMedia.findMany({
-    where: { subjectId: { in: allSubjectIds } },
-    include: {
-      media: {
-        select: {
-          id: true, fileName: true, title: true, mimeType: true,
-          source: { select: { documentType: true } },
-        },
-      },
-    },
-    orderBy: { sortOrder: "asc" },
-    take: 20,
-  });
-  const visibleMedia = studentMedia.filter(
-    (sm) => sm.media.source?.documentType && isStudentVisibleDefault(sm.media.source.documentType),
-  );
-
-  // Build a lookup for filenames (used later in result)
-  const mediaLookup = new Map<string, { fileName: string; title: string | null }>();
-  for (const sm of visibleMedia) {
-    mediaLookup.set(sm.media.id, { fileName: sm.media.fileName, title: sm.media.title });
-  }
-
-  let finalFlowPhases: any = null;
-  if (visibleMedia.length > 0) {
-    const domainRow = await prisma.domain.findUnique({
-      where: { id: domainId },
-      select: { onboardingFlowPhases: true },
-    });
-    const flowConfig = domainRow?.onboardingFlowPhases as { phases?: Array<{ phase: string; duration: string; goals: string[]; content?: Array<{ mediaId: string; instruction?: string }> }> } | null;
-    if (flowConfig?.phases?.length) {
-      // Find the first content-bearing phase by name. Widened from the
-      // previous regex (topic|teach|content|practice|reading) which
-      // failed for domains whose phase names use other vocabulary
-      // (welcome / orient / discover / sample / close — the IELTS
-      // domain), causing the fallback to dump all media into
-      // phase[0] (welcome) with identical placeholder instructions.
-      const contentIdx = flowConfig.phases.findIndex(
-        (p) =>
-          /topic|teach|content|practice|reading|discover|sample|share|present|introduce|explore/i.test(
-            p.phase,
-          ),
-      );
-
-      if (contentIdx < 0) {
-        // No content-bearing phase. Skip media attachment rather than
-        // dumping into phase[0] (typically "welcome" — a greeting,
-        // not a content slot). The operator can attach manually if
-        // they want, and the warning surfaces the missing phase
-        // pattern in production traces.
-        console.warn(
-          `[wizard-tools] create_course: no content phase found in onboarding flow ` +
-            `(phases: ${flowConfig.phases.map((p) => p.phase).join(", ")}); ` +
-            `skipping media attachment for ${visibleMedia.length} item(s).`,
-        );
-      } else {
-        const updatedPhases = flowConfig.phases.map((phase, i) => {
-          if (i !== contentIdx) return phase;
-          return {
-            ...phase,
-            content: visibleMedia.map((sm) => ({
-              mediaId: sm.media.id,
-              // Empty instruction — operator fills in something
-              // specific per media. Previously hardcoded
-              // "Share this with the learner when introducing the topic"
-              // for every row, which produced visually-duplicate UI
-              // rows that confused educators.
-              instruction: "",
-            })),
-          };
-        });
-        finalFlowPhases = { phases: updatedPhases };
-        // #828 — central helper; fans staleness to all playbooks in domain.
-        await updateDomainConfig(
-          domainId,
-          (d) => ({ ...d, onboardingFlowPhases: finalFlowPhases }),
-          { reason: "wizard create_course — onboardingFlowPhases" },
-        );
-      }
-    }
-  }
 
   // 9a. Ensure the wizard user has a TEACHER Caller (needed for educator dashboard + cohort ownership)
   let teacherCaller = await prisma.caller.findFirst({
